+++ src/renderer/services/app/service.ts
@@ -3,8 +3,9 @@ import * as Rx from 'rxjs'
 import { startWith, mapTo, distinctUntilChanged } from 'rxjs/operators'
 import * as RxOp from 'rxjs/operators'
 
-import { Network, SlipTolerance } from '../../../shared/api/types'
+import { Network } from '../../../shared/api/types'
 import { observableState } from '../../helpers/stateHelper'
+import { SlipTolerance } from '../../types/asgardex'
 import { toClientNetwork } from '../clients'
 import { DEFAULT_NETWORK, DEFAULT_SLIP_TOLERANCE } from '../const'
 import { Network$, SlipTolerance$, OnlineStatus } from './types'
